+++ src/validators.js
@@ -1,29 +1,21 @@
 import F from 'futil'
 import _ from 'lodash/fp'
-import { toJS } from 'mobx'
-
-let maybeLimitFields = fields => (fields ? _.pick(fields) : x => x)
+import { toJS } from './mobx'
 
 export let validatorJS = Validator => (form, fields) => {
   let validation = new Validator(
-    form.getSnapshot(),
+    toJS(_.mapValues('value', fields)),
     _.flow(
-      maybeLimitFields(fields),
       _.mapValues(x => toJS(x.rules)),
-      F.compactObject
-    )(form.fields)
+      _.omitBy(_.isEmpty)
+    )(fields)
   )
-  validation.setAttributeNames(_.mapValues('label', form.fields))
+  validation.setAttributeNames(_.mapValues('label', fields))
   return validation.fails() ? validation.errors.all() : {}
 }
 
-export let functions = (form, fields) => {
-  let snapshot = form.getSnapshot()
-  return _.flow(
-    maybeLimitFields(fields),
-    F.mapValuesIndexed(({ validator = () => {} }, field) =>
-      validator(snapshot[field])
-    ),
-    F.compactObject
-  )(form.fields)
-}
+export let functions = (form, fields) =>
+  _.flow(
+    F.mapValuesIndexed(x => (x.validator || _.noop)(x.value, x)),
+    _.omitBy(_.isEmpty)
+  )(fields)